refactor(playground): collapse to single schema-driven implementation
Delete the legacy hand-authored Playground surface (12 entries,
controls module, /playground/<component> routes) and restructure the
schema-driven slice into conventional folders: components/, hooks/,
lib/, entries/. No barrel — app/page.tsx imports the layout directly.

Sidebar ships with Introduction + Button per ADR-0004 §8 until #205
re-authors the remaining Entries. Sweep "toolbar" → "controls sidebar"
through CONTEXT.md, ADR-0004, ADR-0005 to match the implementation.

Document the restructure in ADR-0006: folder shape, walker-lockstep
discipline, parked header-controls placement, deferred nav→real-route
follow-up.
